Fixing PruneList cache

+++ src/Server/Node/Node.cpp
@@ -84,7 +84,7 @@ void Node::UpdateDisplay(const int secondsRunning)
 	}
 	else if (status == ESyncStatus::PROCESSING_TXHASHSET)
 	{
-		std::cout << "\nStatus: Validating TxHashSet (" << pSyncStatus->GetProcessingStatus() << "%)";
+		std::cout << "\nStatus: Validating TxHashSet (" << (int)pSyncStatus->GetProcessingStatus() << "%)";
 	}
 	else if (status == ESyncStatus::TXHASHSET_SYNC_FAILED)
 	{
